requirements.txt
# Core app
Flask>=3.0.0,<4
python-dotenv>=1.0

# Dev tooling
black>=24.8
ruff>=0.6
pre-commit>=3.7
pytest>=8.2
ipykernel>=6.29
